requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.0.0
transformers==4.36.2
torch==2.1.0
reportlab==4.0.7
django-cors-headers==3.14.0
python-dotenv==1.0.0
PyJWT==2.8.0

